+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/debug/renderer/jersey/ConfigOptionsRenderTest.groovy
@@ -39,6 +39,7 @@ class ConfigOptionsRenderTest extends Specification {
 
     Exception mappers
         Throwable                      ExceptionMapperBinder\$1      (i.d.core.setup)
+        Throwable                      DefaultExceptionMapper       (o.g.jersey.server)
         EofException                   EarlyEofExceptionMapper      (i.d.jersey.errors)
         EmptyOptionalException         EmptyOptionalExceptionMapper (i.d.jersey.optional)
         IOException                    GuiceExceptionMapper         (r.v.d.g.c.h.support)
@@ -87,6 +88,7 @@ class ConfigOptionsRenderTest extends Specification {
         byte[]                         ByteArrayProvider            (o.g.j.m.internal)                       [application/octet-stream, */*]
         DataSource                     DataSourceProvider           (o.g.j.m.internal)                       [application/octet-stream, */*]
         Document                       DocumentProvider             (o.g.j.jaxb.internal)                    [application/xml, text/xml, */*]
+        List<EntityPart>               EntityPartReader             (o.g.j.m.m.internal)                     [multipart/form-data]
         Enum                           EnumMessageProvider          (o.g.j.m.internal)                       [text/plain]
         File                           FileProvider                 (o.g.j.m.internal)                       [application/octet-stream, */*]
         MultivaluedMap<String, String> FormMultivaluedMapProvider   (o.g.j.m.internal)                       [application/x-www-form-urlencoded]
@@ -96,8 +98,11 @@ class ConfigOptionsRenderTest extends Specification {
         InputStream                    InputStreamProvider          (o.g.j.m.internal)                       [application/octet-stream, */*]
         Object                         JacksonJsonProvider          (c.f.j.j.rs.json)                        [*/*]
         Object                         JacksonMessageBodyProvider   (i.d.jersey.jackson)                     [*/*]
+        MultiPart                      MultiPartReaderServerSide    (o.g.j.m.m.internal)                     [multipart/*]
+        Path                           PathProvider                 (o.g.j.m.internal)                       [application/octet-stream, */*]
         Reader                         ReaderProvider               (o.g.j.m.internal)                       [text/plain, */*]
         RenderedImage                  RenderedImageProvider        (o.g.j.m.internal)                       [image/*, application/octet-stream]
+        EntityPart                     SingleEntityPartReader       (o.g.j.m.m.internal)
         DOMSource                      DomSourceReader              (o.g.j.m.i.SourceProvider)               [application/xml, text/xml, */*]
         SAXSource                      SaxSourceReader              (o.g.j.m.i.SourceProvider)               [application/xml, text/xml, */*]
         StreamSource                   StreamSourceReader           (o.g.j.m.i.SourceProvider)               [application/xml, text/xml, */*]
@@ -128,6 +133,7 @@ class ConfigOptionsRenderTest extends Specification {
         ChunkedOutput<Object>          ChunkedResponseWriter        (o.g.jersey.server)
         DataSource                     DataSourceProvider           (o.g.j.m.internal)                       [application/octet-stream, */*]
         Document                       DocumentProvider             (o.g.j.jaxb.internal)                    [application/xml, text/xml, */*]
+        List<EntityPart>               EntityPartWriter             (o.g.j.m.m.internal)                     [multipart/form-data]
         Enum                           EnumMessageProvider          (o.g.j.m.internal)                       [text/plain]
         File                           FileProvider                 (o.g.j.m.internal)                       [application/octet-stream, */*]
         MultivaluedMap<String, String> FormMultivaluedMapProvider   (o.g.j.m.internal)                       [application/x-www-form-urlencoded]
@@ -137,13 +143,16 @@ class ConfigOptionsRenderTest extends Specification {
         InputStream                    InputStreamProvider          (o.g.j.m.internal)                       [application/octet-stream, */*]
         Object                         JacksonJsonProvider          (c.f.j.j.rs.json)                        [*/*]
         Object                         JacksonMessageBodyProvider   (i.d.jersey.jackson)                     [*/*]
+        MultiPart                      MultiPartWriter              (o.g.j.m.m.internal)                     [multipart/*]
         OptionalDouble                 OptionalDoubleMessageBodyWriter (i.d.jersey.optional)                 [*/*]
         OptionalInt                    OptionalIntMessageBodyWriter (i.d.jersey.optional)                    [*/*]
         OptionalLong                   OptionalLongMessageBodyWriter (i.d.jersey.optional)                   [*/*]
         Optional<Object>               OptionalMessageBodyWriter    (i.d.jersey.guava)                       [*/*]
         Optional<Object>               OptionalMessageBodyWriter    (i.d.jersey.optional)                    [*/*]
+        Path                           PathProvider                 (o.g.j.m.internal)                       [application/octet-stream, */*]
         Reader                         ReaderProvider               (o.g.j.m.internal)                       [text/plain, */*]
         RenderedImage                  RenderedImageProvider        (o.g.j.m.internal)                       [image/*]
+        EntityPart                     SingleEntityPartWriter       (o.g.j.m.m.internal)
         Source                         SourceWriter                 (o.g.j.m.i.SourceProvider)               [application/xml, text/xml, */*]
         StreamingOutput                StreamingOutputProvider      (o.g.j.m.internal)                       [application/octet-stream, */*]
         String                         StringMessageProvider        (o.g.j.m.internal)                       [text/plain, */*]
@@ -231,6 +240,7 @@ class ConfigOptionsRenderTest extends Specification {
         CookieParamValueParamProvider (o.g.j.s.i.inject)
         DelegatedInjectionValueParamProvider (o.g.j.s.i.inject)
         EntityParamValueParamProvider (o.g.j.s.i.inject)
+        FormDataParamValueParamProvider (o.g.j.m.m.internal)
         FormParamValueParamProvider  (o.g.j.s.i.inject)
         GuiceValueParamProvider      (r.v.d.g.c.h.support)
         HKValueParamProvider         (r.v.d.g.c.h.s.hk)         *jersey managed
@@ -261,6 +271,7 @@ class ConfigOptionsRenderTest extends Specification {
         @QueryParam                     ParamInjectionResolver       (o.g.j.s.i.inject)          using QueryParamValueParamProvider
         @Uri                            ParamInjectionResolver       (o.g.j.s.i.inject)          using WebTargetValueParamProvider
         @BeanParam                      ParamInjectionResolver       (o.g.j.s.i.inject)          using BeanParamValueParamProvider
+        @FormDataParam                  ParamInjectionResolver       (o.g.j.s.i.inject)          using FormDataParamValueParamProvider
 """ as String;
     }