+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/debug/renderer/jersey/PriorityCountInRenderTest.groovy
@@ -45,6 +45,7 @@ class PriorityCountInRenderTest extends Specification {
         IOException                    Mapper1                      (r.v.d.g.d.r.j.PriorityCountInRenderTest)
         IOException                    Mapper2                      (r.v.d.g.d.r.j.PriorityCountInRenderTest)
         Throwable                      ExceptionMapperBinder\$1      (i.d.core.setup)
+        Throwable                      DefaultExceptionMapper       (o.g.jersey.server)
         EofException                   EarlyEofExceptionMapper      (i.d.jersey.errors)
         EmptyOptionalException         EmptyOptionalExceptionMapper (i.d.jersey.optional)
         IllegalStateException          IllegalStateExceptionMapper  (i.d.jersey.errors)